fix(coordinator): migrate import psbt button to its own tsx component
